requirements.txt
streamlit
plotly
requests
beautifulsoup4
pandas
matplotlib
sqlalchemy
schedule
python-telegram-bot
statsmodels
python-dotenv
lxml
openpyxl
xgboost
scikit-learn
yfinance
feedparser
textblob
transformers
torch
pytest
